feature(lotto): 윤혁님 피드백 반영
- ConsoleInput.inputManualLottoNumbers(): lotteryTicketMachine에 대한 의존을 피하기 위해 List<LottoNumber> 를 반환하도록 변경 후 LottoConsole 에서 조립
- LottoConsole: winningTicket.checkTicketAll 을 사용하도록 변경
- LottoConstants: 삭제. TICKET_PRICE 는 LottoTicketMachine 내부의 companion object 에서 관리
- Return -> RateOfReturn 으로 이름 변경
- Money: 내부 값 타입 Long으로 변경(, 연산 관련된 부분도 Long으로 변경)

추가
- ResultView: 연산 기능 없이 화면에 표시해주는 로직만 남기도록 함
- LottoConsole: 수동 티켓 개수를 0개 이하로 입력하면 수동 티켓 번호 입력 건너뛰도록 처리

+++ src/test/kotlin/lotto/MoneyTest.kt
@@ -41,7 +41,7 @@ class MoneyTest : StringSpec({
     "Money는 Int 타입으로 변환할 수 있다" {
         val money1000 = Money(1000)
 
-        money1000.toInt() shouldBe 1000
+        money1000.toLong() shouldBe 1000
     }
 
     "Money는 Double 타입으로 변환할 수 있다" {
